Add simple genetic algorithm
It uses a permutation representation with order crossover and 20%
chance of a single mutation. The fitness function is simply counting
the number of successful rows, columns, and diagonals. The population
is setup as a diffusion model because the generational model converges
way too quickly.

The algorithm fails to find solutions! I believe this is because magic
squares are rare within the search space and because the fitness
function may not be a good indicator of solutions that are approaching
feasibility. It's a good starting point though.

The code depends on cbarrick/evo@da8ff5be, so be sure to update evo.