Fix a representational issue
According to Wikipedia, the sum of every row, column, and diagonal is
known to be m = n * (n^2 + 1) / 2. This is the value used previously to
check the constraints. However, the Wikipedia article discusses magic
squares on the range (0,N] while Evo works with permutations on the
range [0,N). Therefore we adjust the constraint such that the sum
becomes m - n = n * (n^2 - 1) / 2.

This commit also adjusts the String method so that the output is
consistent with the Wikipedia article and the Prolog comparison code.

Now that we're using the proper constraint, the algorithm easily finds
solutions when n=3, but it still fails for n=4.